requirements.txt
pandas>=2.0.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-generativeai>=0.5.0
supabase>=2.3.0
python-dotenv>=1.0.0
tenacity>=8.2.0
